pyproject.toml
[tool.poetry]
name = "photoshare-app"
version = "0.1.0"
description = ""
authors = ["Bohdan Boiprav <122865615+bohdanboiprav@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
aiosqlite = "^0.19.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
